+++ include/zephyr/drivers/eeprom.h
@@ -82,15 +82,6 @@ __subsystem struct eeprom_driver_api {
 __syscall int eeprom_read(const struct device *dev, off_t offset, void *data,
 			  size_t len);
 
-static inline int z_impl_eeprom_read(const struct device *dev, off_t offset,
-				     void *data, size_t len)
-{
-	const struct eeprom_driver_api *api =
-		(const struct eeprom_driver_api *)dev->api;
-
-	return api->read(dev, offset, data, len);
-}
-
 /**
  *  @brief Write data to EEPROM
  *
@@ -105,15 +96,6 @@ __syscall int eeprom_write(const struct device *dev, off_t offset,
 			   const void *data,
 			   size_t len);
 
-static inline int z_impl_eeprom_write(const struct device *dev, off_t offset,
-				      const void *data, size_t len)
-{
-	const struct eeprom_driver_api *api =
-		(const struct eeprom_driver_api *)dev->api;
-
-	return api->write(dev, offset, data, len);
-}
-
 /**
  *  @brief Get the size of the EEPROM in bytes
  *
@@ -123,14 +105,6 @@ static inline int z_impl_eeprom_write(const struct device *dev, off_t offset,
  */
 __syscall size_t eeprom_get_size(const struct device *dev);
 
-static inline size_t z_impl_eeprom_get_size(const struct device *dev)
-{
-	const struct eeprom_driver_api *api =
-		(const struct eeprom_driver_api *)dev->api;
-
-	return api->size(dev);
-}
-
 #ifdef __cplusplus
 }
 #endif
@@ -139,6 +113,7 @@ static inline size_t z_impl_eeprom_get_size(const struct device *dev)
  * @}
  */
 
+#include <zephyr/drivers/eeprom/internal/eeprom_impl.h>
 #include <zephyr/syscalls/eeprom.h>
 
 #endif /* ZEPHYR_INCLUDE_DRIVERS_EEPROM_H_ */
